Change curly quotes to straight quotes
Thank you for this post! It was exactly what I was looking for right now and it helped me play with GitHub Actions for the first time. The only thing that got me going through your article was that I blankly copied and pasted your `eleventy_build.yml` code, and it contained curly quotes instead of straight quotes. After I realized this, I simply changed them and it worked perfectly. So hopefully this change will help the next person!